tools/requirements.txt
requests
colorama
pillow
matplotlib
numpy
sklearn